MTDSA-12900 PR comments
